(LeitnerSRS): fix EASY interval as int, not float

+++ src/lib/LeitnerSRS.php
@@ -199,7 +199,7 @@ class LeitnerSRS
       // easy answers get a higher interval
       if ($answer === FlashcardReview::RATE_EASY)
       {
-        $card_interval = ceil($card_interval * self::EASY_FACTOR);
+        $card_interval = (int) ceil($card_interval * self::EASY_FACTOR);
       }
 
       // add variance to spread due cards so that they don't all fall onto the same days
